fix: applies recommended changes
