+++ cyclops-ctrl/pkg/cluster/k8sclient/client.go
@@ -18,7 +18,6 @@ import (
 	"k8s.io/apimachinery/pkg/watch"
 
 	"github.com/cyclops-ui/cyclops/cyclops-ctrl/api/v1alpha1"
-	"github.com/gin-gonic/gin"
 
 	"gopkg.in/yaml.v2"
 
@@ -171,7 +170,7 @@ func (k *KubernetesClient) GetPods(namespace, name string) ([]apiv1.Pod, error) 
 	return podList.Items, err
 }
 
-func (k *KubernetesClient) GetStreamedPodLogs(ctx *gin.Context, namespace, container, name string, logCount *int64, logChan chan<- string) error {
+func (k *KubernetesClient) GetStreamedPodLogs(ctx context.Context, namespace, container, name string, logCount *int64, logChan chan<- string) error {
 	podLogOptions := apiv1.PodLogOptions{
 		Container:  container,
 		TailLines:  logCount,
@@ -180,7 +179,7 @@ func (k *KubernetesClient) GetStreamedPodLogs(ctx *gin.Context, namespace, conta
 	}
 
 	podClient := k.clientset.CoreV1().Pods(namespace).GetLogs(name, &podLogOptions)
-	stream, err := podClient.Stream(ctx.Request.Context())
+	stream, err := podClient.Stream(ctx)
 	if err != nil {
 		return err
 	}
